+++ src/main/kotlin/study/Language.kt
@@ -1,0 +1,3 @@
+package study
+
+class Language(val name: String, val level: Int)
