+++ src/main/kotlin/study/Skill.kt
@@ -1,0 +1,3 @@
+package study
+
+class Skill(val skillType: SkillType, val message: String)
