+++ src/main/kotlin/study/SkillType.kt
@@ -1,0 +1,12 @@
+package study
+
+enum class SkillType {
+    SOFT, HARD;
+
+    private lateinit var value: String
+
+    fun value(value: String): SkillType {
+        this.value = value
+        return this
+    }
+}
