+++ src/test/kotlin/study/DslTest.kt
@@ -104,77 +104,3 @@ class DslTest {
         person.languages[1].level shouldBe 3
     }
 }
-
-fun introduce(block: Person.() -> Unit): Person = Person().apply(block)
-
-class Person {
-    lateinit var name: String
-    lateinit var company: String
-    lateinit var skills: List<Skill>
-    lateinit var languages: List<Language>
-
-    fun name(value: String) {
-        name = value
-    }
-
-    fun company(value: String) {
-        company = value
-    }
-
-    fun skills(block: SkillsBuilder.() -> Unit) = SkillsBuilder()
-        .apply(block).build().let { skills = it }
-
-    fun languages(block: LanguagesBuilder.() -> Unit) = LanguagesBuilder().apply(block).build().let { languages = it }
-}
-
-class LanguagesBuilder {
-    private val languagesList = mutableListOf<Language>()
-
-    infix fun String.level(level: Int) {
-        languagesList.add(Language(this, level))
-    }
-
-    fun build(): List<Language> = languagesList
-}
-
-class SkillsBuilder {
-    private val skillsList = mutableListOf<Skill>()
-
-    fun soft(value: String) {
-        skillsList.add(
-            Skill().apply {
-                skillType = SkillType.SOFT
-                message = value
-            }
-        )
-    }
-
-    fun hard(value: String) {
-        skillsList.add(
-            Skill().apply {
-                skillType = SkillType.HARD
-                message = value
-            }
-        )
-    }
-
-    fun build(): List<Skill> = skillsList
-}
-
-class Skill {
-    lateinit var skillType: SkillType
-    lateinit var message: String
-}
-
-class Language(val name: String, val level: Int)
-
-enum class SkillType {
-    SOFT, HARD;
-
-    private lateinit var value: String
-
-    fun value(value: String): SkillType {
-        this.value = value
-        return this
-    }
-}
